requirements.txt
torch
torchaudio
numpy
scipy
geoopt
fastapi
uvicorn
mlflow
dvc
python-osc
pylive
sounddevice
pysoundfile
